+++ src/enums/confirmation_method.rs
@@ -1,6 +1,6 @@
 use num_enum::{IntoPrimitive, TryFromPrimitive};
 use serde_repr::{Deserialize_repr, Serialize_repr};
-use strum_macros::{Display, EnumString};
+use strum::{Display, EnumString};
 
 /// The method of confirmation.
 #[derive(Debug, Serialize_repr, Deserialize_repr, Display, EnumString, PartialEq, TryFromPrimitive, IntoPrimitive, Clone, Copy)]