+++ src/enums/offer_filter.rs
@@ -1,6 +1,6 @@
 use num_enum::{IntoPrimitive, TryFromPrimitive};
 use serde_repr::{Deserialize_repr, Serialize_repr};
-use strum_macros::{Display, EnumString};
+use strum::{Display, EnumString};
 
 /// Filter for getting trade offers.
 #[derive(Serialize_repr, Deserialize_repr, Display, EnumString, Debug, PartialEq, TryFromPrimitive, IntoPrimitive, Clone)]